Enhanced price history endpoint: structured JSON output with formatted timestamps. Fixed data type for price fields to float due to SQLite limitations (no decimal support).

+++ database/migrations/2025_02_12_114638_create_products_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('products', function (Blueprint $table) {
             $table->id();
             $table->string('name')->unique();
-            $table->decimal('price', 10, 2);
+            $table->float('price', 10, 2);
             $table->integer('stock');
             $table->timestamps();
         });
